+++ alkymi/recipes.py
@@ -1,9 +1,8 @@
 from pathlib import Path
-from typing import List, Tuple, Optional, Any, Dict, Iterable, Union
+from typing import List, Tuple, Any, Dict, Iterable, Union, TypeVar, Type, Optional
 
 from .config import CacheType
 from .recipe import Recipe
-from .types import Outputs
 
 
 def glob_files(name: str, directory: Path, pattern: str, recursive: bool, cache=CacheType.Auto) -> Recipe[List[Path]]:
@@ -30,7 +29,7 @@ def glob_files(name: str, directory: Path, pattern: str, recursive: bool, cache=
         else:
             return list(directory.glob(pattern))
 
-    def _check_clean(last_outputs: Optional[Outputs]) -> bool:
+    def _check_clean(last_outputs: List[Path]) -> bool:
         """
         If rerunning glob produces the same list of files, then the recipe is clean
 
@@ -39,7 +38,7 @@ def glob_files(name: str, directory: Path, pattern: str, recursive: bool, cache=
         """
         if last_outputs is None:
             return False
-        return _glob_recipe() == last_outputs[0]
+        return _glob_recipe() == last_outputs
 
     return Recipe(_glob_recipe, [], name, transient=False, cache=cache, cleanliness_func=_check_clean)
 
@@ -108,119 +107,79 @@ def zip_results(name: str, recipes: Iterable[Recipe], cache=CacheType.Auto) \
     return Recipe(_zip_results, recipes, name, transient=False, cache=cache)
 
 
-class NamedArgs(Recipe[Dict[Any, Any]]):
+T = TypeVar('T')
+
+
+class Arg(Recipe[T]):
     """
-    Class providing stateful keyword argument(s)
+    Class providing a stateful argument
 
-    To use, create a NamedArgs instance with the initial value for your arguments, e.g. ``NamedArgs(val0=0, val1=1,
-    val2=2)``, then provide as a recipe to downstream recipes. To change the input arguments, call ``set_args()`` again
-    - this will mark the contained recipe as dirty and cause reevaluation of downstream recipes
-    """
-
-    def __init__(self, name: str, cache=CacheType.Auto, **_kwargs: Any):
-        """
-        Create a new NamedArgs instance with initial argument value(s)
-
-        :param name: The name to give the created Recipe
-        :param cache: The type of caching to use for this Recipe
-        :param _kwargs: The initial keyword argument value(s)
-        """
-        self._kwargs = _kwargs  # type: Dict[Any, Any]
-        super().__init__(self._produce_kwargs, [], name, transient=False, cache=cache, cleanliness_func=self._clean)
-
-    def _produce_kwargs(self) -> Dict[Any, Any]:
-        """
-        :return: The current set of keyword arguments
-        """
-        return self._kwargs
-
-    def _clean(self, last_outputs: Optional[Outputs]) -> bool:
-        """
-        Checks whether the arguments have changed since the last evaluation
-
-        :param last_outputs: The last set of outputs (arguments)
-        :return: True if the arguments remain the same
-        """
-        if last_outputs is None:
-            return self._kwargs is None
-        return self._kwargs == last_outputs[0]
-
-    def set_args(self, **_kwargs) -> None:
-        """
-        Change the arguments, causing the recipe to need reevaluation
-
-        :param _kwargs: The new set of keyword arguments
-        """
-        # TODO(mathias): Consider enforcing argument count and types here
-        self._kwargs = _kwargs
-
-
-class Args(Recipe[Tuple[Any, ...]]):
-    """
-    Class providing stateful non-keyword argument(s)
-
-    To use, create an Args instance with the initial value for your arguments, e.g. ``Args(0, 1, 2)``, then provide as a
-    recipe to downstream recipes. To change the input arguments, call ``set_args()`` again - this will mark
-    the recipe as dirty and cause reevaluation of downstream recipe(s)
+    To use, create an Arg instance with the initial value for your argument, e.g. ``Arg(0)``, then provide as a
+    recipe to downstream recipes. To change the input arguments, call ``set()`` again - this will mark the recipe as
+    dirty and cause reevaluation of downstream recipe(s)
     """
 
-    def __init__(self, *_args: Any, name: str, cache=CacheType.Auto):
+    def __init__(self, arg: T, name: str, cache=CacheType.Auto):
         """
-        Create a new Args instance with initial argument value(s)
+        Create a new Arg instance with initial argument value
 
-        :param _args: The initial argument value(s)
+        :param arg: The initial argument value
         :param name: The name to give the created Recipe
         :param cache: The type of caching to use for this Recipe
         """
-        self._args = _args  # type: Tuple[Any, ...]
-        super().__init__(self._produce_args, [], name, transient=False, cache=cache, cleanliness_func=self._clean)
+        self._arg = arg
+        self._type = type(arg)
+        self._subtype = next((type(val) for val in arg), None) if isinstance(arg, Iterable) else None
+        super().__init__(self._produce_arg, [], name, transient=False, cache=cache, cleanliness_func=self._clean)
 
-    def _produce_args(self) -> Tuple[Any, ...]:
+    @property
+    def type(self) -> Type[T]:
         """
-        :return: The current set of arguments
+        :return: The type of the argument
         """
-        return self._args
+        return self._type
 
-    def _clean(self, last_outputs: Optional[Outputs]) -> bool:
+    @property
+    def subtype(self) -> Optional[Any]:
         """
-        Checks whether the arguments have changed since the last evaluation
-
-        :param last_outputs: The last set of outputs (arguments)
-        :return: True if the arguments remain the same
+        :return: The subtype of the argument (e.g. the type of items contained in a list). Will be None for non-iterable
+                 types
         """
-        if last_outputs is None:
-            return self._args is None
-        return self._args == last_outputs
+        return self._subtype
 
-    def set_args(self, *_args) -> None:
+    def _produce_arg(self) -> T:
         """
-        Change the arguments, causing the recipe to need reevaluation
-
-        :param _args: The new set of arguments
+        :return: The current argument
         """
-        # TODO(mathias): Consider enforcing argument count and types here
-        self._args = _args
+        return self._arg
+
+    def _clean(self, last_outputs: Any) -> bool:
+        """
+        Checks whether the argument has changed since the last evaluation
+
+        :param last_outputs: The last set of outputs (argument)
+        :return: True if the argument remain the same
+        """
+        return self._arg == last_outputs
+
+    def set(self, _arg: T) -> None:
+        """
+        Change the argument, causing the recipe to need reevaluation
+
+        :param _arg: The new argument
+        """
+        if not isinstance(_arg, self._type):
+            raise TypeError("Type of argument ({}) not {}".format(type(_arg), self._type))
+        self._arg = _arg
 
 
-def kwargs(name: str, cache=CacheType.Auto, **_kwargs: Any) -> NamedArgs:
+def arg(_arg: T, name: str, cache=CacheType.Auto) -> Arg[T]:
     """
-    Shorthand for creating a ``NamedArgs`` instance
+    Shorthand for creating an ``Arg`` instance
 
+    :param _arg: The initial argument to use
     :param name: The name to give the created Recipe
     :param cache: The type of caching to use for this Recipe
-    :param _kwargs: The initial keyword arguments to use
-    :return: The created ``NamedArgs`` instance
+    :return: The created ``Arg`` instance
     """
-    return NamedArgs(name, cache, **_kwargs)
-
-
-def args(*_args: Any, name: str, cache=CacheType.Auto) -> Args:
-    """
-    Shorthand for creating an ``Args`` instance
-
-    :param _args: The initial arguments to use
-    :param name: The name to give the created Recipe
-    :param cache: The type of caching to use for this Recipe
-    :return: The created ``Args`` instance
-    """
-    return Args(*_args, name=name, cache=cache)
+    return Arg(_arg, name=name, cache=cache)